% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages}
\alias{packages}
\title{Conveniently load and install packages}
\usage{
packages(..., repos = "http://cran.r-project.org", dep = TRUE,
  lib.loc = NULL, quietly = FALSE, warn.conflicts = TRUE)
}
\arguments{
\item{...}{comma seperated package names and/or GitHub repository address.
In case of the latter, ensure that you enter the address as a character in the format \code{username/repo} (or \code{username/repo/subdir} if necessary).}

\item{repos}{URL of the repository to use, see \code{\link{install.packages}}.}

\item{dep}{logical indicating whether to also install packages which the requested packages depend on / link to / import / suggest. See \code{\link{install.packages}}.}

\item{lib.loc}{character vector describing the location of \code{R} library trees to search for, or NULL. See \code{\link{require}}.}

\item{quietly}{logical indicating whether warning / error messages from \code{\link{install.packages}}, \code{\link{install_github}}, and \code{\link{require}} should be printed.}

\item{warn.conflicts}{logical indicating whether \code{\link{require}} warnings about conflicts from attaching a new package should be printed.}
}
\description{
\code{packages} loads and, if necessary, installs packages.
Adapted from the \href{https://github.com/russellpierce/repsych/blob/master/R/glibrary.R}{glibrary} function in the repsych package.
Packages can be installed from both CRAN and GitHub.
}
\examples{
packages(ggplot2) # Install and load from CRAN
packages(ggplot2, "hrbrmstr/hrbrthemes") # Load packages from CRAN and GitHub in one line

}
\seealso{
\code{\link{install.packages}}, \code{\link[devtools]{install_github}}, \code{\link{require}}
}
\author{
Frank H. Hezemans, \email{Frank.Hezemans@mrc-cbu.cam.ac.uk}
}
