% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedDiffHist.R
\name{pairedDiffHist}
\alias{pairedDiffHist}
\title{Add histogram of paired difference on the identity line}
\usage{
pairedDiffHist(p, origin, maxheight, colour = "black", fill = "grey",
  size = 1, bins = NULL, binwidth = NULL, meanline = FALSE,
  meanline_pos = 0, meanline_colour = "black", meanline_size = 1,
  meanline_type = "solid", sigstars = NULL, sigstars_pos = 0,
  sigstars_colour = "black", sigstars_size = 1)
}
\arguments{
\item{p}{the ggplot object to which the histogram should be added.}

\item{origin}{number indicating the center of the histogram on the plot.}

\item{maxheight}{number indicating the desired height of the tallest bin.}

\item{colour, fill, size}{aesthetics for the histogram bins.}

\item{bins, binwidth}{the number of bins / the width of the bins.}

\item{meanline}{logical. Should a line segment at the mean of the histogram be added?}

\item{meanline_pos}{number indicating by how much the line should extend the dimensions of the biggest histogram bin.}

\item{meanline_colour, meanline_size, meanline_type}{aesthetics of the mean line.}

\item{sigstars}{number indicating the amount of stars to be added at the mean of the distribution, to indicate significance level of difference. Default is NULL.}

\item{sigstars_pos}{number indicating by how much the significance stars should be shifted up from the top of the mean line}

\item{sigstars_colour, sigstars_size}{aesthetics of the significance stars.}
}
\description{
\code{pairedDiffHist} Adds a histogram on the identity line which illustrates the distribution of the difference between two conditions.
This is a useful addition to a paired scatterplot, to illustrate how the conditions differ within the sample.
Each histogram bin is manually added to the plot with \code{geom_polygon}.
}
\details{
In order for the histogram to make sense, the axes of the plot should be identically scaled. Ideally the aspect ratio is set to one, using \code{theme(aspect.ratio = 1)}.
}
\examples{
library(ggplot2)
library(UsingR) # Package contains dataset

# Get data
data <- crime

# Create paired scatterplot
paired_scatter <- ggplot(data = data, aes(x = y1983, y = y1993)) +
    geom_abline(slope = 1, intercept = 0) +
    geom_point(shape = 21, fill = "grey") +
    xlim(0, 1500) + ylim(0, 1500) +
    labs(title = "Violent crime in 50 U.S. states in 1983 and 1993",
         subtitle = "Rates per 100,000 population",
         x = "Violent crime rate in 1983",
         y = "Violent crime rate in 1993") +
    theme_minimal() +
    theme(aspect.ratio = 1)

# Add histogram of paired differences
paired_scatter <- pairedDiffHist(paired_scatter, origin = 1250, maxheight = 100,
                                 colour = "black", fill = "grey", size = 1,
                                 bins = 6, meanline = TRUE, sigstars = 3,
                                 sigstars_pos = 25)

}
\seealso{
\code{\link[ggplot2]{geom_histogram}}, \code{\link[ggplot2]{geom_polygon}}, \code{\link[ggplot2]{annotate}}
}
\author{
Frank H. Hezemans, \email{Frank.Hezemans@mrc-cbu.cam.ac.uk}
}
