% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileLines.R
\name{quantileLines}
\alias{quantileLines}
\title{Add quantile lines to a paired scatterplot}
\usage{
quantileLines(p, probs = c(0.25, 0.5, 0.75), colour_m = "darkgrey",
  colour = "grey", size_m = 0.5, size = 0.25, linetype_m = "dashed",
  linetype = "dashed")
}
\arguments{
\item{p}{the ggplot object to which the quantile lines should be added.}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{colour_m}{character indicating the colour of the middle line.}

\item{colour}{character indicating the colour of the other lines.}

\item{size_m}{numeric indicating the thickness of the middle line.}

\item{size}{numeric indicating the thickness of the other lines.}

\item{linetype_m}{character indicating the linetype of the middle line.}

\item{linetype}{character indicating the linetype of the other lines.}
}
\description{
\code{quantileLines} Adds line segments at the requested quantiles of the x and y variables.
This is a useful addition to a paired scatterplot, to illustrate how observations are linked between conditions.
The quantile lines are set as the lowest layer, so they won't obscure the original paired scatterplot.
Adapted from the \href{https://github.com/GRousselet/rogme/blob/master/R/plot_utils.R}{plot_quartile_bars} function in the rogme package.
}
\details{
In order for these quantile lines to make sense, the axes of the plot should be identically scaled.
}
\examples{
library(ggplot2)
library(UsingR) # Package contains dataset

# Get data
data <- crime

# Create paired scatterplot
paired_scatter <- ggplot(data = data, aes(x = y1983, y = y1993)) +
    geom_abline(slope = 1, intercept = 0) +
    geom_point(shape = 21, fill = "grey") +
    xlim(0, 1500) + ylim(0, 1500) +
    labs(title = "Violent crime in 50 U.S. states in 1983 and 1993",
         subtitle = "Rates per 100,000 population",
         x = "Violent crime rate in 1983",
         y = "Violent crime rate in 1993") +
    theme_minimal() +
    theme(aspect.ratio = 1)

# Add lines at the first quartile, median, and third quartile
paired_scatter <- quantileLines(paired_scatter, probs = c(0.25, 0.5, 0.75),
                                colour_m = "black", colour = "grey",
                                size_m = 0.5, size = 0.25,
                                linetype_m = "longdash", linetype = "dashed")          

}
\seealso{
\code{\link[stats]{quantile}}, \code{\link[ggplot2]{geom_segment}}
}
\author{
Frank H. Hezemans, \email{Frank.Hezemans@mrc-cbu.cam.ac.uk}
}
